# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "WebServices",
        "version": "1.0.0"
    },
    "paths": {
...
        "/stocks/{id}": {
            "get": {
                "tags": [
                    "Stock"
                ],
                "summary": "Return a stock",
                "operationId": "App\\Http\\Controllers\\StockController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "stock id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "get stock",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Stock"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Stock": {
                "properties": {
                    "product_id": {
                        "schema": "StockCollection",
                        "type": "id",
                        "example": "product_id"
                    },
                    "shop_id": {
                        "schema": "StockCollection",
                        "type": "id",
                        "example": "shop_id"
                    },
                    "price": {
                        "schema": "StockCollection",
                        "type": "integer",
                        "example": "price"
                    },
                    "stock": {
                        "schema": "StockCollection",
                        "type": "integer",
                        "example": "stock"
                    },
                    "sales": {
                        "schema": "StockCollection",
                        "type": "integer",
                        "example": "sales"
                    }
                },
                "type": "object"
            }
        }
    },
    "security": [
        []
    ]
}